+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next Release
 ### Features
 - Add progress UI for registering a patient
+- Add progress UI when scheduling appointment in `ScheduleAppointmentSheet`
 
 ###Internal
 - Migrated `LoginPinScreen` to Mobius